Updated usage comment in voglib_host.h
There was incorrect usage information I had written before the
interface was implemented that needed to be updated.
I forgot to include that vog_init() required the name of the
program to be written to the device, vog_run() needed to be called
seperate of vog_init(), and I ended up naming my read/write
functions vog_read() and vog_write() instead of the initially
planned vog_get() and vog_put().