Wrote basic test and fixed various errors
Wrote a basic test file intended to take a character, write to
the client, then read the response from the client, and print it
out. It doesn't work yet, so there's more bug-hunting to be done.
Compiling the test revieled various errors, so fixed what I have
found thus far.